rework(general): define DAA standard at build with new CONFIG_NAVIGATOR_ADSB_F3442
