+++ .github/workflows/checks.yml
@@ -77,6 +77,7 @@ jobs:
           PX4_SBOM_DISABLE: 1
         run: |
           make tests
+          make tests_daa_crosstrack
           make tests_vtest_moving
 
       - uses: ./.github/actions/save-ccache
