+++ boards/px4/fmu-v6c/visionTargetEstStatic.px4board
@@ -82,6 +82,7 @@ CONFIG_SYSTEMCMDS_VER=y
 CONFIG_SYSTEMCMDS_WORK_QUEUE=y
 
 CONFIG_MAVLINK_DIALECT="development"
+CONFIG_NAVIGATOR_ADSB_F3442=y
 CONFIG_MODULES_VISION_TARGET_ESTIMATOR=y
 CONFIG_MODULES_LANDING_TARGET_ESTIMATOR=n
 CONFIG_MODULES_ROVER_POS_CONTROL=n
