+++ boards/px4/sitl/default.px4board
@@ -43,6 +43,7 @@ CONFIG_MODULES_MC_POS_CONTROL=y
 CONFIG_MODULES_MC_RATE_CONTROL=y
 CONFIG_MODULES_NAVIGATOR=y
 CONFIG_MODE_NAVIGATOR_VTOL_TAKEOFF=y
+CONFIG_NAVIGATOR_ADSB_F3442=y
 CONFIG_NAVIGATOR_ADSB_FAKE_TRAFFIC=y
 CONFIG_NUM_MISSION_ITMES_SUPPORTED=10000
 CONFIG_MODULES_PAYLOAD_DELIVERER=y
