+++ boards/px4/sitl/test-daa-crosstrack.px4board
@@ -1,0 +1,3 @@
+CONFIG_BOARD_NOLOCKSTEP=y
+CONFIG_DRIVERS_DISTANCE_SENSOR_LIGHTWARE_LASER_SERIAL=y
+CONFIG_NAVIGATOR_ADSB_F3442=n
