+++ docs/en/msg_docs/DetectAndAvoid.md
@@ -41,8 +41,6 @@ Name | Type | Value | Description
 <a id="#DAA_CONFLICT_LVL_MEDIUM"></a> DAA_CONFLICT_LVL_MEDIUM | `uint8` | 2 |
 <a id="#DAA_CONFLICT_LVL_HIGH"></a> DAA_CONFLICT_LVL_HIGH | `uint8` | 3 |
 <a id="#DAA_CONFLICT_LVL_CRITICAL"></a> DAA_CONFLICT_LVL_CRITICAL | `uint8` | 4 |
-<a id="#DAA_STANDARD_CROSSTRACK"></a> DAA_STANDARD_CROSSTRACK | `uint8` | 0 |
-<a id="#DAA_STANDARD_F3442"></a> DAA_STANDARD_F3442 | `uint8` | 1 | ASTM F3442/F3442M-based conflict evaluation
 <a id="#UNIQUE_ID_ENCODING_ICAO"></a> UNIQUE_ID_ENCODING_ICAO | `uint8` | 0 | `unique_id` contains an ICAO address
 <a id="#UNIQUE_ID_ENCODING_ADSB_CALLSIGN"></a> UNIQUE_ID_ENCODING_ADSB_CALLSIGN | `uint8` | 1 | `unique_id` contains an ADS-B callsign packed into a `uint64`
 <a id="#UNIQUE_ID_ENCODING_UAS_ID"></a> UNIQUE_ID_ENCODING_UAS_ID | `uint8` | 2 | `unique_id` contains the reduced tail bytes of a UAS ID
@@ -73,9 +71,6 @@ uint8 DAA_CONFLICT_LVL_LOW = 1
 uint8 DAA_CONFLICT_LVL_MEDIUM = 2
 uint8 DAA_CONFLICT_LVL_HIGH = 3
 uint8 DAA_CONFLICT_LVL_CRITICAL = 4
-
-uint8 DAA_STANDARD_CROSSTRACK = 0
-uint8 DAA_STANDARD_F3442 = 1 # ASTM F3442/F3442M-based conflict evaluation
 
 uint8 UNIQUE_ID_ENCODING_ICAO = 0 # `unique_id` contains an ICAO address
 uint8 UNIQUE_ID_ENCODING_ADSB_CALLSIGN = 1 # `unique_id` contains an ADS-B callsign packed into a `uint64`
