+++ src/modules/navigator/DetectAndAvoid/detect_and_avoid_params.yaml
@@ -26,195 +26,3 @@ parameters:
       min: 1
       max: 120
       reboot_required: true
-    DAA_STANDARD:
-      description:
-        short: Standard used for DAA
-        long: |-
-          Select the detect-and-avoid standard.
-
-          0: use Crosstrack. This preserves the legacy upstream traffic avoidance behavior and uses NAV_TRAFF_ parameters.
-          1: use the F3442 standard and configure F34_ parameters. This supports four conflict levels by checking four volumes.
-      type: enum
-      values:
-        0: Crosstrack
-        1: F3442
-      default: 1
-      min: 0
-      max: 1
-      reboot_required: true
-    DAA_DFLT_VEL:
-      description:
-        short: Default vertical speed of other aircraft
-        long: |-
-          The vertical speed reported by transponder modules cannot always be trusted.
-          If DAA_EN_DFLT_VEL is enabled, overwrite the other aircraft's reported vertical velocity and use DAA_DFLT_VEL.
-      type: float
-      default: 10.0
-      unit: m/s
-      min: 0
-      reboot_required: true
-    DAA_EN_DFLT_VEL:
-      description:
-        short: Ignore reported vertical speed of other aircraft
-        long: If enabled, ignore the other aircraft's reported vertical velocity and use DAA_DFLT_VEL.
-      type: boolean
-      default: 1
-      reboot_required: true
-    DAA_LVL_LOW_ACT:
-      description:
-        short: Low conflict level action
-        long: |-
-          Trigger an automated action if the low conflict level is breached.
-          Low conflict level boundaries are defined by the configured DAA standard.
-          Changing this parameter at runtime does not re-evaluate conflicts that are already buffered.
-          The updated action is only applied on later conflict transitions, and automatic mode changes are only issued on conflict escalation.
-      type: enum
-      values:
-        0: Disabled
-        1: Warn only
-        2: Return mode
-        3: Land mode
-        4: Position Hold mode
-        5: Terminate
-      default: 1
-      min: 0
-      max: 5
-    DAA_LVL_MED_ACT:
-      description:
-        short: Medium conflict level action
-        long: |-
-          Trigger an automated action if the medium conflict level is breached.
-          Medium conflict level boundaries are defined by the configured DAA standard.
-          Changing this parameter at runtime does not re-evaluate conflicts that are already buffered.
-          The updated action is only applied on later conflict transitions, and automatic mode changes are only issued on conflict escalation.
-      type: enum
-      values:
-        0: Disabled
-        1: Warn only
-        2: Return mode
-        3: Land mode
-        4: Position Hold mode
-        5: Terminate
-      default: 1
-      min: 0
-      max: 5
-    DAA_LVL_HIGH_ACT:
-      description:
-        short: High conflict level action
-        long: |-
-          Trigger an automated action if the high conflict level is breached.
-          High conflict level boundaries are defined by the configured DAA standard.
-          Changing this parameter at runtime does not re-evaluate conflicts that are already buffered.
-          The updated action is only applied on later conflict transitions, and automatic mode changes are only issued on conflict escalation.
-      type: enum
-      values:
-        0: Disabled
-        1: Warn only
-        2: Return mode
-        3: Land mode
-        4: Position Hold mode
-        5: Terminate
-      default: 1
-      min: 0
-      max: 5
-    DAA_LVL_CRIT_ACT:
-      description:
-        short: Critical conflict level action
-        long: |-
-          Trigger an automated action if the critical conflict level is breached.
-          Critical conflict level boundaries are defined by the configured DAA standard.
-          Changing this parameter at runtime does not re-evaluate conflicts that are already buffered.
-          The updated action is only applied on later conflict transitions, and automatic mode changes are only issued on conflict escalation.
-      type: enum
-      values:
-        0: Disabled
-        1: Warn only
-        2: Return mode
-        3: Land mode
-        4: Position Hold mode
-        5: Terminate
-      default: 1
-      min: 0
-      max: 5
-    F34_LVL_CRIT_RAD:
-      description:
-        short: F3442 critical conflict cylinder radius
-        long: |-
-          In the F3442 standard, the critical conflict level corresponds to the Near Mid-Air Collision (NMAC) volume.
-          If two aircraft NMAC zones overlap, the action defined in DAA_LVL_CRIT_ACT is triggered.
-          In the F3442 standard, an NMAC occurs when two aircraft come within 100 ft (31 m) vertically and 500 ft (153 m) horizontally of each other while in flight.
-      type: float
-      default: 77.0
-      unit: m
-      min: 10
-      reboot_required: true
-    F34_LVL_CRIT_HGT:
-      description:
-        short: F3442 critical conflict cylinder height
-        long: |-
-          In the F3442 standard, the critical conflict level corresponds to the Near Mid-Air Collision (NMAC) volume.
-          If two aircraft NMAC zones overlap, the action defined in DAA_LVL_CRIT_ACT is triggered.
-          In the F3442 standard, an NMAC occurs when two aircraft come within 100 ft (31 m) vertically and 500 ft (153 m) horizontally of each other while in flight.
-      type: float
-      default: 16.0
-      unit: m
-      min: 10
-      reboot_required: true
-    F34_LVL_HIGH_RAD:
-      description:
-        short: F3442 high conflict cylinder radius
-        long: |-
-          In the F3442 standard, the high conflict level corresponds to the Well Clear (WC) volume.
-          If two aircraft WC zones overlap, the action defined in DAA_LVL_HIGH_ACT is triggered.
-          In the F3442 standard, a WC occurs when two aircraft come within 2000 ft (77 m) vertically and 6250 ft (610 m) horizontally of each other while in flight.
-      type: float
-      default: 305.0
-      unit: m
-      min: 10
-      reboot_required: true
-    F34_LVL_HIGH_HGT:
-      description:
-        short: F3442 high conflict cylinder height
-        long: |-
-          In the F3442 standard, the high conflict level corresponds to the Well Clear (WC) volume.
-          If two aircraft WC zones overlap, the action defined in DAA_LVL_HIGH_ACT is triggered.
-          In the F3442 standard, a WC occurs when two aircraft come within 2000 ft (77 m) vertically and 6250 ft (610 m) horizontally of each other while in flight.
-      type: float
-      default: 39.0
-      unit: m
-      min: 10
-      reboot_required: true
-    F34_LVL_MED_TIME:
-      description:
-        short: Latency time for the medium conflict cylinder
-        long: |-
-          The latency time is designed to ensure the DAA system has time to avoid an NMAC breach.
-          It is used to calculate the medium conflict cylinder:
-          ```
-          radius = F34_LVL_CRIT_RAD + aircraft speed * F34_LVL_MED_TIME
-          height = F34_LVL_CRIT_HGT + aircraft speed * F34_LVL_MED_TIME
-          ```
-
-          If two aircraft zones overlap, the action defined in DAA_LVL_MED_ACT is triggered.
-      type: int32
-      default: 33
-      unit: s
-      min: 0
-      reboot_required: true
-    F34_LVL_LOW_TIME:
-      description:
-        short: Latency time for the low conflict cylinder
-        long: |-
-          The latency time is designed to ensure the DAA system has time to avoid a WC breach.
-          It is used to calculate the low conflict cylinder:
-          ```
-          radius = F34_LVL_HIGH_RAD + aircraft speed * F34_LVL_LOW_TIME
-          height = F34_LVL_HIGH_HGT + aircraft speed * F34_LVL_LOW_TIME
-          ```
-
-          If two aircraft zones overlap, the action defined in DAA_LVL_LOW_ACT is triggered.
-      type: int32
-      default: 33
-      unit: s
-      min: 0
-      reboot_required: true
