+++ src/modules/navigator/DetectAndAvoid/test/detect_and_avoid_f3442_test.cpp
@@ -31,595 +31,30 @@
  *
  ****************************************************************************/
 
-#include <gtest/gtest.h>
-#include <cstring>
-#include <limits>
-#include <memory>
-#include <string>
-#include <vector>
+/**
+ * @file detect_and_avoid_f3442_test.cpp
+ * @brief Tests specific to the ASTM F3442 DAA standard.
+ *
+ * This file is only built when CONFIG_NAVIGATOR_ADSB_F3442 is enabled; it tests the
+ * four nested alert volumes, the per-level actions, and the default velocity handling
+ * that only exist in F3442 builds.
+ *
+ * @author Jonas Perolini <jonspero@me.com>
+ *
+ */
 
-#include "detect_and_avoid.h"
-#include "../navigator.h"
-#include "../mission.h"
-#include <px4_platform_common/px4_work_queue/WorkQueueManager.hpp>
-#include <uORB/topics/mavlink_log.h>
+#include "detect_and_avoid_test_common.h"
 
-using namespace time_literals;
-
-extern "C" __EXPORT int dataman_main(int argc, char *argv[]);
-
-/* Set simple bounds */
-static constexpr float kDfltF34LvlCriticalRad = 150.f;
-static constexpr float kDfltF34LvlCriticalHgt = 30.f;
-static constexpr float kDfltF34LvlHighRad = 600.f;
-static constexpr float kDfltF34LvlHighHgt = 80.f;
-static constexpr int kDfltF34LvlMediumTime = 30.f;
-static constexpr int kDfltF34LvlLowTime = 30.f;
-
-static constexpr float kDfltDaaDfltVel = 5.f;
-static constexpr int kDfltDaaEnDfltVel = 1;
-static constexpr int kDfltF3442Standard = detect_and_avoid_s::DAA_STANDARD_F3442;
-static constexpr uint8_t kDaaConflictLvlNone = 0;
-static constexpr uint8_t kDaaConflictLvlLow = 1;
-static constexpr uint8_t kDaaConflictLvlMedium = 2;
-static constexpr uint8_t kDaaConflictLvlHigh = 3;
-static constexpr uint8_t kDaaConflictLvlCritical = 4;
-
-static constexpr int action_param_value(const DaaAction action)
+// WHY: Activation must fail closed when an F3442 alert-volume parameter is invalid.
+// WHAT: Inject negative volume radii and verify the module refuses to activate.
+TEST_F(DetectAndAvoidTest, ActivationFailsWithInvalidF3442Params)
 {
-	switch (action) {
-	case DaaAction::kDisabled:
-		return 0;
-
-	case DaaAction::kWarnOnly:
-		return 1;
-
-	case DaaAction::kReturnMode:
-		return 2;
-
-	case DaaAction::kLandMode:
-		return 3;
-
-	case DaaAction::kPositionHoldMode:
-		return 4;
-
-	case DaaAction::kTerminate:
-		return 5;
-
-	default:
-		return 0;
-	}
-}
-
-/* Default drone position. Not important as traffic is defined relatively to uav */
-static constexpr float kDfltLatUav = 150.f;
-static constexpr float kDfltLonUav = 30.f;
-static constexpr float kDfltAltUav = 600.f;
-static constexpr hrt_abstime kOrbWaitTimeoutUs = 200000;
-static constexpr useconds_t kOrbWaitPollIntervalUs = 1000;
-
-// Start dataman once for the lifetime of the test binary. Each Navigator constructs
-// multiple DatamanClient instances and each one waits up to 1s for a client ID, so
-// restarting dataman between test suites adds ~4 s of timeout per parameterized case.
-// Same pattern as in src/modules/navigator/test/test_mission_base.cpp.
-class NavigatorDatamanRuntime
-{
-public:
-	NavigatorDatamanRuntime()
-	{
-		param_control_autosave(false);
-		px4::WorkQueueManagerStart();
-
-		char name[] = "dataman";
-		char start[] = "start";
-		char ram[] = "-r";
-		char *argv[] = {name, start, ram};
-		dataman_main(3, argv);
-	}
-
-	~NavigatorDatamanRuntime()
-	{
-		param_control_autosave(true);
-
-		char name[] = "dataman";
-		char stop[] = "stop";
-		char *argv[] = {name, stop};
-		dataman_main(2, argv);
-
-		px4::WorkQueueManagerStop();
-	}
-};
-
-static NavigatorDatamanRuntime &navigatorDatamanRuntime()
-{
-	static NavigatorDatamanRuntime runtime{};
-	return runtime;
-}
-
-// To run: make tests TESTFILTER=detect_and_avoid
-class DetectAndAvoidTest : public ::testing::Test
-{
-
-protected:
-
-	DetectAndAvoidTest() {}
-	~DetectAndAvoidTest() override {}
-
-	void SetUp() override
-	{
-		navigatorDatamanRuntime();
-		param_reset_all();
-		set_DFLT_daa_params(); // NMAC: (150, 30), WC: (600, 80), Aug-WC: 30s, Aug-NMAC: 30s
-		navigator = std::make_unique<Navigator>();
-	}
-
-	void TearDown() override
-	{
-		navigator.reset();
-		transponder_report_s tr{};
-		_traffic_pub.publish(tr);
-	}
-
-public:
-
-	uORB::SubscriptionData<detect_and_avoid_s> _detect_and_avoid_sub{ORB_ID(detect_and_avoid)};
-	uORB::SubscriptionData<detect_and_avoid_most_urgent_s> _detect_and_avoid_most_urgent_sub{ORB_ID(detect_and_avoid_most_urgent)};
-	uORB::SubscriptionData<vehicle_command_s> _vehicle_command_sub{ORB_ID(vehicle_command)};
-	uORB::SubscriptionData<transponder_report_s> _traffic_sub{ORB_ID(transponder_report)};
-	uORB::Subscription _mavlink_log_sub{ORB_ID(mavlink_log)};
-	uORB::Subscription _global_pos_sub{ORB_ID(vehicle_global_position)};
-	uORB::Subscription _vehicle_status_state_sub{ORB_ID(vehicle_status)};
-	uORB::Subscription _vehicle_local_position_sub{ORB_ID(vehicle_local_position)};
-	uORB::Subscription _vehicle_land_detected_sub{ORB_ID(vehicle_land_detected)};
-
-	uORB::Publication<transponder_report_s> _traffic_pub{ORB_ID(transponder_report)};
-	uORB::Publication<vehicle_global_position_s>		_global_pos_pub{ORB_ID(vehicle_global_position)};
-	uORB::Publication<vehicle_status_s>		_vehicle_status_pub{ORB_ID(vehicle_status)};
-	uORB::Publication<vehicle_local_position_s>		_vehicle_local_position_pub{ORB_ID(vehicle_local_position)};
-	uORB::Publication<vehicle_land_detected_s>		_vehicle_land_detected_pub{ORB_ID(vehicle_land_detected)};
-	uORB::Publication<parameter_update_s>		_parameter_update_pub{ORB_ID(parameter_update)};
-	std::unique_ptr<Navigator> navigator{};
-
-	struct breach_distances_s {
-		float nmac_hor;
-		float nmac_vert;
-		float wc_hor;
-		float wc_vert;
-		float aug_nmac_hor;
-		float aug_nmac_vert;
-		float aug_wc_hor;
-		float aug_wc_vert;
-	};
-
-	void get_DFLT_breach_distances(const matrix::Vector3f uav_vel, const float traffic_hor_vel,
-				       breach_distances_s &breach_distance)
-	{
-
-		const float uav_hor_vel = uav_vel.xy().norm();
-
-		/* NMAC */
-		breach_distance.nmac_hor = 2 * kDfltF34LvlCriticalRad;
-		breach_distance.nmac_vert = 2 * kDfltF34LvlCriticalHgt;
-
-		/* WC */
-		breach_distance.wc_hor = 2 * kDfltF34LvlHighRad;
-		breach_distance.wc_vert = 2 * kDfltF34LvlHighHgt;
-
-		/* Aug NMAC */
-		breach_distance.aug_nmac_hor = 2 * kDfltF34LvlCriticalRad + kDfltF34LvlMediumTime * (fabsf(
-						       traffic_hor_vel) + fabsf(uav_hor_vel));
-		breach_distance.aug_nmac_vert = 2 * kDfltF34LvlCriticalHgt + kDfltF34LvlMediumTime * (fabsf(
-							kDfltDaaDfltVel) + fabsf(uav_vel(2)));
-
-		/* Aug WC */
-		breach_distance.aug_wc_hor = 2 * kDfltF34LvlHighRad + kDfltF34LvlLowTime * (fabsf(
-						     traffic_hor_vel) + fabsf(uav_hor_vel));
-		breach_distance.aug_wc_vert = 2 * kDfltF34LvlHighHgt + kDfltF34LvlLowTime * (fabsf(
-						      kDfltDaaDfltVel) + fabsf(uav_vel(2)));
-	}
-
-	void publish_local_pos_vel(const matrix::Vector3f local_pos_vel, const hrt_abstime timestamp = hrt_absolute_time(),
-				   const float heading = 0.f)
-	{
-		vehicle_local_position_s l_pos{};
-		l_pos.timestamp_sample = timestamp;
-		l_pos.timestamp = timestamp;
-		l_pos.vx = local_pos_vel(0);
-		l_pos.vy = local_pos_vel(1);
-		l_pos.vz = local_pos_vel(2);
-		l_pos.heading = heading;
-		_vehicle_local_position_pub.publish(l_pos);
-	}
-
-	void publish_global_pos(const double lat, const double lon, const double alt, const hrt_abstime timestamp = hrt_absolute_time())
-	{
-		vehicle_global_position_s g_pos{};
-		g_pos.timestamp_sample = timestamp;
-		g_pos.lat = lat;
-		g_pos.lon = lon;
-		g_pos.alt = alt;
-		g_pos.timestamp = timestamp;
-		_global_pos_pub.publish(g_pos);
-	}
-
-	void publish_vehicle_status(const uint8_t nav_state, const uint8_t arming_state)
-	{
-
-		vehicle_status_s v_status{};
-		v_status.timestamp = hrt_absolute_time();
-		v_status.nav_state = nav_state;
-		v_status.arming_state = arming_state;
-		_vehicle_status_pub.publish(v_status);
-	}
-
-	void publish_land_status(const bool has_landed)
-	{
-
-		vehicle_land_detected_s vehicle_land{};
-		vehicle_land.timestamp = hrt_absolute_time();
-		vehicle_land.landed = has_landed;
-		_vehicle_land_detected_pub.publish(vehicle_land);
-
-	}
-
-	void publish_parameter_update()
-	{
-		parameter_update_s update{};
-		update.timestamp = hrt_absolute_time();
-		_parameter_update_pub.publish(update);
-	}
-
-	transponder_report_s create_transponder_report(const uint32_t icao_address, const char *callsign, const double lat,
-			const double lon, const float altitude, const float hor_velocity, const float ver_velocity,
-			const uint16_t flags)
-	{
-		transponder_report_s tr{};
-		tr.timestamp = hrt_absolute_time();
-		tr.icao_address = icao_address;
-		tr.lat = lat;
-		tr.lon = lon;
-		tr.altitude = altitude;
-		tr.heading = 0.f;
-		tr.hor_velocity = hor_velocity;
-		tr.ver_velocity = ver_velocity;
-		tr.altitude_type = 0;
-		tr.emitter_type = 1;
-		tr.tslc = 1;
-		tr.flags = flags;
-		strncpy(&tr.callsign[0], callsign, sizeof(tr.callsign) - 1);
-		tr.callsign[sizeof(tr.callsign) - 1] = '\0';
-		return tr;
-	}
-
-	void publish_transponder_report(const transponder_report_s &tr)
-	{
-		_traffic_pub.publish(tr);
-	}
-
-	template<typename Predicate>
-	void wait_until(const Predicate &predicate, const hrt_abstime timeout_us = kOrbWaitTimeoutUs)
-	{
-		const hrt_abstime start = hrt_absolute_time();
-		bool condition_met = predicate();
-
-		while (!condition_met && hrt_elapsed_time(&start) < timeout_us) {
-			px4_usleep(kOrbWaitPollIntervalUs);
-			condition_met = predicate();
-		}
-
-		ASSERT_TRUE(condition_met);
-	}
-
-	void wait_for_topic_update(uORB::Subscription &subscription, const hrt_abstime timeout_us = kOrbWaitTimeoutUs)
-	{
-		wait_until([&subscription]() {
-			return subscription.updated();
-		}, timeout_us);
-	}
-
-	void drain_transponder_report_topic()
-	{
-		while (_traffic_sub.update()) {}
-	}
-
-	template<typename Publisher>
-	void publish_traffic_and_check(const Publisher &publisher)
-	{
-		drain_transponder_report_topic();
-		publisher();
-		wait_for_topic_update(_traffic_sub);
-		navigator->check_traffic();
-	}
-
-	void publish_transponder_report_and_check(const transponder_report_s &tr)
-	{
-		publish_traffic_and_check([&]() {
-			publish_transponder_report(tr);
-		});
-	}
-
-	void set_DFLT_daa_params()
-	{
-		param_set(param_handle(px4::params::F34_LVL_CRIT_RAD), &kDfltF34LvlCriticalRad);
-		param_set(param_handle(px4::params::F34_LVL_CRIT_HGT), &kDfltF34LvlCriticalHgt);
-		param_set(param_handle(px4::params::F34_LVL_HIGH_RAD), &kDfltF34LvlHighRad);
-		param_set(param_handle(px4::params::F34_LVL_HIGH_HGT), &kDfltF34LvlHighHgt);
-		param_set(param_handle(px4::params::F34_LVL_MED_TIME), &kDfltF34LvlMediumTime);
-		param_set(param_handle(px4::params::F34_LVL_LOW_TIME), &kDfltF34LvlLowTime);
-
-		param_set(param_handle(px4::params::DAA_EN_DFLT_VEL), &kDfltDaaEnDfltVel);
-		param_set(param_handle(px4::params::DAA_DFLT_VEL), &kDfltDaaDfltVel);
-		param_set(param_handle(px4::params::DAA_STANDARD), &kDfltF3442Standard);
-
-	}
-
-	void set_default_uav_state(const double lat_uav, const double lon_uav, const float alt_uav,
-				   const matrix::Vector3f uav_vel)
-	{
-
-		/* Navigation state */
-		const bool has_landed = false;
-		const uint8_t arming_state = vehicle_status_s::ARMING_STATE_ARMED;
-		const uint8_t nav_state = vehicle_status_s::NAVIGATION_STATE_AUTO_MISSION;
-
-		/* Publish UAV state */
-		publish_global_pos(lat_uav, lon_uav, alt_uav);
-		publish_land_status(has_landed);
-		publish_vehicle_status(nav_state, arming_state);
-		publish_local_pos_vel(uav_vel);
-	}
-
-	void generate_random_callsign(char *callsign, size_t length = 8)
-	{
-		const char *chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-		const size_t chars_length = 36; // Length of the chars array.
-
-		for (size_t i = 0; i < length; ++i) {
-			const size_t index = rand() % chars_length;
-			callsign[i] = chars[index];
-		}
-
-		callsign[length] = '\0'; // Null-terminate the string.
-	}
-
-	void generate_random_UAS_ID(uint8_t uas_id[PX4_GUID_BYTE_LENGTH])
-	{
-		for (int i = 0; i < PX4_GUID_BYTE_LENGTH; i++) {
-			uas_id[i] = rand() % 256;
-		}
-	}
-
-	/* Test conflict priorities */
-	const uint8_t states_enable_hold [4] {
-		vehicle_status_s::NAVIGATION_STATE_AUTO_MISSION,
-		vehicle_status_s::NAVIGATION_STATE_AUTO_TAKEOFF,
-		vehicle_status_s::NAVIGATION_STATE_AUTO_FOLLOW_TARGET,
-		vehicle_status_s::NAVIGATION_STATE_AUTO_VTOL_TAKEOFF,
-	};
-
-	const uint8_t states_enable_rtl [2] {
-		vehicle_status_s::NAVIGATION_STATE_ORBIT,
-		vehicle_status_s::NAVIGATION_STATE_AUTO_LOITER,
-	};
-
-	const uint8_t states_enable_land [1] {
-		vehicle_status_s::NAVIGATION_STATE_AUTO_RTL
-	};
-
-	const uint8_t states_enable_termination [8] {
-		vehicle_status_s::NAVIGATION_STATE_AUTO_LAND,
-		vehicle_status_s::NAVIGATION_STATE_DESCEND,
-		vehicle_status_s::NAVIGATION_STATE_AUTO_PRECLAND,
-		vehicle_status_s::NAVIGATION_STATE_MANUAL,
-		vehicle_status_s::NAVIGATION_STATE_ALTCTL,
-		vehicle_status_s::NAVIGATION_STATE_POSCTL,
-		vehicle_status_s::NAVIGATION_STATE_ACRO,
-		vehicle_status_s::NAVIGATION_STATE_STAB
-	};
-
-	const uint8_t states_never_enable [2] {
-		vehicle_status_s::NAVIGATION_STATE_TERMINATION,
-		vehicle_status_s::NAVIGATION_STATE_OFFBOARD,
-	};
-
-	struct expected_outcome_s {
-		bool hold_allowed;
-		bool rtl_allowed;
-		bool land_allowed;
-		bool terminate_allowed;
-	};
-
-	void check_eval_conflict_escalation_action(Navigator *navigator_instance, const expected_outcome_s &expected_outcome,
-			DaaAction action)
-	{
-
-		const uint8_t arming_state = vehicle_status_s::ARMING_STATE_ARMED;
-
-		for (const uint8_t nav_state : states_enable_hold) {
-			publish_vehicle_status(nav_state, arming_state);
-			sync_navigator_topics(navigator_instance);
-			ASSERT_TRUE(navigator_instance->get_detect_and_avoid()->eval_conflict_escalation_action(action) ==
-				    expected_outcome.hold_allowed);
-		}
-
-		for (const uint8_t nav_state : states_enable_rtl) {
-			publish_vehicle_status(nav_state, arming_state);
-			sync_navigator_topics(navigator_instance);
-			ASSERT_TRUE(navigator_instance->get_detect_and_avoid()->eval_conflict_escalation_action(action) ==
-				    expected_outcome.rtl_allowed);
-		}
-
-		for (const uint8_t nav_state : states_enable_land) {
-			publish_vehicle_status(nav_state, arming_state);
-			sync_navigator_topics(navigator_instance);
-			ASSERT_TRUE(navigator_instance->get_detect_and_avoid()->eval_conflict_escalation_action(action) ==
-				    expected_outcome.land_allowed);
-		}
-
-		for (const uint8_t nav_state : states_enable_termination) {
-			publish_vehicle_status(nav_state, arming_state);
-			sync_navigator_topics(navigator_instance);
-			ASSERT_TRUE(navigator_instance->get_detect_and_avoid()->eval_conflict_escalation_action(action) ==
-				    expected_outcome.terminate_allowed);
-		}
-
-		for (const uint8_t nav_state : states_never_enable) {
-			publish_vehicle_status(nav_state, arming_state);
-			sync_navigator_topics(navigator_instance);
-			ASSERT_FALSE(navigator_instance->get_detect_and_avoid()->eval_conflict_escalation_action(action));
-		}
-	};
-
-	void check_highest_conflict(const uint8_t conflict_level, const bool action_required)
-	{
-		if (_detect_and_avoid_most_urgent_sub.updated()) {
-			ASSERT_TRUE(_detect_and_avoid_most_urgent_sub.update());
-		}
-
-		const detect_and_avoid_most_urgent_s &daa_status = _detect_and_avoid_most_urgent_sub.get();
-		ASSERT_TRUE(daa_status.conflict_level == conflict_level);
-		ASSERT_TRUE(daa_status.has_action == action_required);
-	}
-
-	void expect_empty_most_urgent_status(const detect_and_avoid_most_urgent_s &status)
-	{
-		EXPECT_EQ(status.conflict_level, kDaaConflictLvlNone);
-		EXPECT_FALSE(status.has_action);
-		EXPECT_EQ(status.unique_id, 0ULL);
-		EXPECT_EQ(status.unique_id_encoding, detect_and_avoid_most_urgent_s::UNIQUE_ID_ENCODING_ICAO);
-		EXPECT_FLOAT_EQ(status.aircraft_dist, 9999.f);
-		EXPECT_GT(status.timestamp, 0u);
-	}
-
-	std::vector<std::string> drain_mavlink_logs()
-	{
-		std::vector<std::string> logs;
-		mavlink_log_s mavlink_log{};
-
-		while (_mavlink_log_sub.update(&mavlink_log)) {
-			logs.emplace_back(mavlink_log.text);
-		}
-
-		return logs;
-	}
-
-	size_t count_logs_with_prefix(const std::vector<std::string> &logs, const char *prefix)
-	{
-		size_t count = 0;
-
-		for (const std::string &log : logs) {
-			if (log.rfind(prefix, 0) == 0) {
-				++count;
-			}
-		}
-
-		return count;
-	}
-
-	bool any_log_contains(const std::vector<std::string> &logs, const char *needle)
-	{
-		for (const std::string &log : logs) {
-			if (log.find(needle) != std::string::npos) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	void drain_detect_and_avoid_topic()
-	{
-		while (_detect_and_avoid_sub.update()) {}
-	}
-
-	void drain_detect_and_avoid_most_urgent_topic()
-	{
-		while (_detect_and_avoid_most_urgent_sub.update()) {}
-	}
-
-	void recreate_navigator()
-	{
-		navigator.reset();
-		navigator = std::make_unique<Navigator>();
-	}
-
-	void sync_navigator_topics(Navigator *navigator_instance = nullptr)
-	{
-		Navigator *target = navigator_instance ? navigator_instance : navigator.get();
-
-		if (_global_pos_sub.updated()) {
-			vehicle_global_position_s global_position{};
-
-			if (_global_pos_sub.copy(&global_position)) {
-				*target->get_global_position() = global_position;
-			}
-		}
-
-		if (_vehicle_land_detected_sub.updated()) {
-			vehicle_land_detected_s land_detected{};
-
-			if (_vehicle_land_detected_sub.copy(&land_detected)) {
-				*target->get_land_detected() = land_detected;
-			}
-		}
-
-		vehicle_local_position_s local_position{};
-
-		if (_vehicle_local_position_sub.copy(&local_position)) {
-			*target->get_local_position() = local_position;
-		}
-
-		vehicle_status_s vehicle_status{};
-
-		if (_vehicle_status_state_sub.copy(&vehicle_status)) {
-			*target->get_vstatus() = vehicle_status;
-		}
-	}
-};
-
-struct action_priority_test_case_s {
-	const char *name;
-	DaaAction action;
-	DetectAndAvoidTest::expected_outcome_s expected_outcome;
-};
-
-class DetectAndAvoidActionPriorityTest : public DetectAndAvoidTest, public ::testing::WithParamInterface<action_priority_test_case_s>
-{
-};
-
-std::string action_priority_test_case_name(const ::testing::TestParamInfo<action_priority_test_case_s> &info)
-{
-	return info.param.name;
-}
-
-// WHY: Activation and deactivation gate all DetectAndAvoid processing and must fail closed on bad configuration.
-// WHAT: Verify successful startup with default params, then check disabled and invalid parameter sets leave the module inactive.
-TEST_F(DetectAndAvoidTest, OnActivation)
-{
-	// GIVEN: DetectAndAvoid starts with the default valid parameter set.
+	// GIVEN: The module activated cleanly with the default parameter set.
 	EXPECT_TRUE(navigator->get_detect_and_avoid()->is_activated());
-
-	// WHEN: The module is explicitly deactivated.
-	navigator->get_detect_and_avoid()->on_inactivation();
-
-	// THEN: It reports itself inactive.
-	EXPECT_FALSE(navigator->get_detect_and_avoid()->is_activated());
-
-	// WHEN: The feature is disabled through parameters before activation.
-	const int daa_en = 0;
-	param_t param_daa_en = param_handle(px4::params::DAA_EN);
-	param_set(param_daa_en, &daa_en);
-
-	navigator->get_detect_and_avoid()->on_activation();
-
-	// THEN: Activation is refused while the feature is disabled.
-	EXPECT_FALSE(navigator->get_detect_and_avoid()->is_activated());
-	param_reset_all();
 
 	// WHEN: A required F3442 parameter is invalid.
 	const float negative_value = -10;
-	param_t param_f34_lvl_critical_rad = param_handle(px4::params::F34_LVL_CRIT_RAD);
-	param_set(param_f34_lvl_critical_rad, &negative_value);
+	param_set(param_handle(px4::params::F34_LVL_CRIT_RAD), &negative_value);
 
 	navigator->get_detect_and_avoid()->on_activation();
 
@@ -627,23 +62,10 @@ TEST_F(DetectAndAvoidTest, OnActivation)
 	EXPECT_FALSE(navigator->get_detect_and_avoid()->is_activated());
 	param_reset_all();
 
-	param_t param_f34_lvl_high_rad = param_handle(px4::params::F34_LVL_HIGH_RAD);
-	param_set(param_f34_lvl_high_rad, &negative_value);
+	param_set(param_handle(px4::params::F34_LVL_HIGH_RAD), &negative_value);
 
 	navigator->get_detect_and_avoid()->on_activation();
 	EXPECT_FALSE(navigator->get_detect_and_avoid()->is_activated());
-	param_reset_all();
-
-	// WHEN: An unknown DAA standard is selected.
-	const uint8_t daa_standard = 5;
-	param_t param_daa_standard = param_handle(px4::params::DAA_STANDARD);
-	param_set(param_daa_standard, &daa_standard);
-
-	navigator->get_detect_and_avoid()->on_activation();
-
-	// THEN: Unsupported standards are rejected as well.
-	EXPECT_FALSE(navigator->get_detect_and_avoid()->is_activated());
-
 	param_reset_all();
 }
 
@@ -695,7 +117,7 @@ TEST_F(DetectAndAvoidTest, OnInactivationRepublishesClearedMostUrgentState)
 }
 
 // WHY: Failed activation should clear the published most-urgent state so downstream consumers never keep stale traffic after DAA stops.
-// WHAT: Start from a published critical conflict, force activation failure with an invalid standard, and verify the replacement most-urgent sample is cleared.
+// WHAT: Start from a published critical conflict, force activation failure with invalid F3442 parameters, and verify the replacement most-urgent sample is cleared.
 TEST_F(DetectAndAvoidTest, FailedActivationPublishesBenignNoConflictOutput)
 {
 	// GIVEN: Ownship is initialized and DetectAndAvoid has already published a critical most-urgent conflict.
@@ -728,10 +150,10 @@ TEST_F(DetectAndAvoidTest, FailedActivationPublishesBenignNoConflictOutput)
 	EXPECT_EQ(_detect_and_avoid_most_urgent_sub.get().conflict_level, kDaaConflictLvlCritical);
 	drain_detect_and_avoid_most_urgent_topic();
 
-	const uint8_t invalid_standard = 5;
-	param_set(param_handle(px4::params::DAA_STANDARD), &invalid_standard);
+	const float invalid_critical_radius = -10.f;
+	param_set(param_handle(px4::params::F34_LVL_CRIT_RAD), &invalid_critical_radius);
 
-	// WHEN: Activation is retried with an invalid DAA standard.
+	// WHEN: Activation is retried with invalid F3442 parameters.
 	navigator->get_detect_and_avoid()->on_activation();
 
 	wait_until([&]() {
@@ -742,19 +164,6 @@ TEST_F(DetectAndAvoidTest, FailedActivationPublishesBenignNoConflictOutput)
 	EXPECT_FALSE(navigator->get_detect_and_avoid()->is_activated());
 	ASSERT_TRUE(_detect_and_avoid_most_urgent_sub.update());
 	expect_empty_most_urgent_status(_detect_and_avoid_most_urgent_sub.get());
-}
-
-// WHY: ICAO identifiers are vehicle-specific, so the default configuration must not suppress unrelated traffic.
-// WHAT: Read the default primary and secondary ICAO parameters and verify both are disabled by default.
-TEST_F(DetectAndAvoidTest, PrimaryAndSecondaryIcaoDefaultsDisabled)
-{
-	int32_t primary_icao = 0;
-	ASSERT_EQ(param_get(param_handle(px4::params::ADSB_ICAO_ID), &primary_icao), 0);
-	EXPECT_EQ(primary_icao, -1);
-
-	int32_t secondary_icao = 0;
-	ASSERT_EQ(param_get(param_handle(px4::params::ADSB_ICAO_ID_2), &secondary_icao), 0);
-	EXPECT_EQ(secondary_icao, -1);
 }
 
 // WHY: The optional default vertical traffic speed directly changes F3442 conflict classification.
@@ -1135,34 +544,6 @@ TEST_F(DetectAndAvoidTest, ProcessesAllQueuedTrafficReports)
 	EXPECT_EQ(conflict.conflict_level, kDaaConflictLvlCritical);
 }
 
-// WHY: Operators need a way to cancel a queued fake-traffic script before it publishes more synthetic traffic.
-// WHAT: Arm a fake-traffic mode, stop it before the next navigator cycle, and verify no synthetic conflicts are emitted.
-TEST_F(DetectAndAvoidTest, StopFakeTrafficCancelsPendingScript)
-{
-	const double lat_uav = 46.52342;
-	const double lon_uav = 6.524234;
-	const float alt_uav = 400.f;
-	const matrix::Vector3f uav_vel{5.f, 10.f, 2.f};
-
-	set_default_uav_state(lat_uav, lon_uav, alt_uav, uav_vel);
-	sync_navigator_topics();
-
-	navigator->run_fake_traffic(DetectAndAvoid::FakeTraffMode::kQueueFill);
-	navigator->stop_fake_traffic();
-	drain_mavlink_logs();
-
-	navigator->check_traffic();
-
-	conflict_info_s conflict;
-	navigator->get_detect_and_avoid()->get_most_urgent_conflict(conflict);
-
-	EXPECT_EQ(conflict.conflict_level, kDaaConflictLvlNone);
-
-	const std::vector<std::string> logs = drain_mavlink_logs();
-	EXPECT_EQ(count_logs_with_prefix(logs, "DAA New and Main:"), 0u);
-	EXPECT_EQ(count_logs_with_prefix(logs, "DAA New "), 0u);
-}
-
 // WHY: Zero latitude is a valid coordinate and must not be rejected as an uninitialized value.
 // WHAT: Publish traffic on the equator and verify it still produces a conflict.
 TEST_F(DetectAndAvoidTest, AcceptsTrafficOnZeroLatitude)
@@ -1407,184 +788,6 @@ TEST_F(DetectAndAvoidTest, RuntimeF3442ActionUpdateAppliesOnNextEscalationOnly)
 	EXPECT_TRUE(_detect_and_avoid_most_urgent_sub.get().has_action);
 	ASSERT_TRUE(_vehicle_command_sub.update());
 	EXPECT_EQ(_vehicle_command_sub.get().command, vehicle_command_s::VEHICLE_CMD_DO_SET_MODE);
-}
-
-// WHY: Runtime NAV_TRAFF_AVOID action updates must not re-evaluate the current buffer, but later conflict escalations must use the refreshed action value.
-// WHAT: Start with a crosstrack conflict in warn-only mode, switch NAV_TRAFF_AVOID to terminate at runtime, verify no immediate command is sent, then clear and re-trigger the conflict and confirm termination is commanded.
-TEST_F(DetectAndAvoidTest, RuntimeCrosstrackTerminateUpdateAppliesOnNextConflict)
-{
-	const int crosstrack_standard = detect_and_avoid_s::DAA_STANDARD_CROSSTRACK;
-	const int warn_only = action_param_value(DaaAction::kWarnOnly);
-	const int terminate_action = action_param_value(DaaAction::kTerminate);
-	param_set(param_handle(px4::params::DAA_STANDARD), &crosstrack_standard);
-	param_set(param_handle(px4::params::NAV_TRAFF_AVOID), &warn_only);
-	recreate_navigator();
-
-	const double lat_uav = 46.52342;
-	const double lon_uav = 6.524234;
-	const float alt_uav = 400.f;
-	const matrix::Vector3f uav_vel{0.f, 0.f, 0.f};
-	const float conflict_distance = 200.f;
-	const float resolve_distance = 5000.f;
-	const float traffic_heading = 3.f * M_PI_2_F;
-	const float traffic_direction = M_PI_2_F;
-	const float hor_velocity = 30.f;
-
-	set_default_uav_state(lat_uav, lon_uav, alt_uav, uav_vel);
-	sync_navigator_topics();
-	drain_detect_and_avoid_most_urgent_topic();
-
-	while (_vehicle_command_sub.update()) {}
-
-	const auto publish_crosstrack_sample = [&](const float distance) {
-		publish_traffic_and_check([&]() {
-			navigator->get_detect_and_avoid()->fake_traffic(14545057, "DDF0A1", distance, traffic_direction, traffic_heading,
-					0.f, hor_velocity, 0.f, 1, lat_uav, lon_uav, alt_uav);
-		});
-	};
-
-	publish_crosstrack_sample(conflict_distance);
-
-	ASSERT_TRUE(_detect_and_avoid_most_urgent_sub.update());
-	EXPECT_EQ(_detect_and_avoid_most_urgent_sub.get().conflict_level, kDaaConflictLvlHigh);
-	EXPECT_FALSE(_detect_and_avoid_most_urgent_sub.get().has_action);
-	EXPECT_FALSE(_vehicle_command_sub.update());
-
-	param_set(param_handle(px4::params::NAV_TRAFF_AVOID), &terminate_action);
-	publish_parameter_update();
-	drain_detect_and_avoid_most_urgent_topic();
-
-	while (_vehicle_command_sub.update()) {}
-
-	navigator->check_traffic();
-	EXPECT_FALSE(_detect_and_avoid_most_urgent_sub.update());
-	EXPECT_FALSE(_vehicle_command_sub.update());
-
-	publish_crosstrack_sample(resolve_distance);
-
-	ASSERT_TRUE(_detect_and_avoid_most_urgent_sub.update());
-	EXPECT_EQ(_detect_and_avoid_most_urgent_sub.get().conflict_level, kDaaConflictLvlNone);
-	EXPECT_FALSE(_detect_and_avoid_most_urgent_sub.get().has_action);
-
-	publish_crosstrack_sample(conflict_distance);
-
-	ASSERT_TRUE(_detect_and_avoid_most_urgent_sub.update());
-	EXPECT_EQ(_detect_and_avoid_most_urgent_sub.get().conflict_level, kDaaConflictLvlHigh);
-	EXPECT_TRUE(_detect_and_avoid_most_urgent_sub.get().has_action);
-	ASSERT_TRUE(_vehicle_command_sub.update());
-	EXPECT_EQ(_vehicle_command_sub.get().command, vehicle_command_s::VEHICLE_CMD_DO_FLIGHTTERMINATION);
-}
-
-// WHY: Invalid traffic altitude should be rejected before any DAA math runs to avoid propagating NaNs through conflict evaluation.
-// WHAT: Publish a report with a NaN altitude and verify that no detect_and_avoid sample is published and no conflict is buffered.
-TEST_F(DetectAndAvoidTest, RejectsNonFiniteTrafficAltitude)
-{
-	// GIVEN: Ownship is initialized and the detect_and_avoid topic is drained.
-	const double lat_uav = 46.52342;
-	const double lon_uav = 6.524234;
-	const float alt_uav = 400.f;
-	const matrix::Vector3f uav_vel{5.f, 10.f, 2.f};
-
-	set_default_uav_state(lat_uav, lon_uav, alt_uav, uav_vel);
-	sync_navigator_topics();
-	drain_detect_and_avoid_topic();
-
-	const uint16_t flags = transponder_report_s::PX4_ADSB_FLAGS_VALID_COORDS |
-			       transponder_report_s::PX4_ADSB_FLAGS_VALID_ALTITUDE;
-	transponder_report_s tr = create_transponder_report(14545057, "DDF0A1", lat_uav, lon_uav,
-				  std::numeric_limits<float>::quiet_NaN(), 0.f, 0.f, flags);
-
-	// WHEN: Traffic is published with a NaN altitude.
-	publish_transponder_report_and_check(tr);
-
-	// THEN: No output is published and no conflict is buffered.
-	EXPECT_FALSE(_detect_and_avoid_sub.update());
-
-	conflict_info_s conflict;
-	navigator->get_detect_and_avoid()->get_most_urgent_conflict(conflict);
-	EXPECT_EQ(conflict.conflict_level, kDaaConflictLvlNone);
-}
-
-// WHY: Crosstrack mode depends on a finite traffic heading and must fail closed if the heading value is NaN or Inf.
-// WHAT: Switch to crosstrack mode, publish invalid heading values with the VALID_HEADING flag set, and verify no output or buffered conflict is produced.
-TEST_F(DetectAndAvoidTest, CrosstrackRejectsNonFiniteTrafficHeading)
-{
-	// GIVEN: DetectAndAvoid is recreated in crosstrack mode.
-	const int crosstrack_standard = detect_and_avoid_s::DAA_STANDARD_CROSSTRACK;
-	param_set(param_handle(px4::params::DAA_STANDARD), &crosstrack_standard);
-	recreate_navigator();
-
-	// GIVEN: Ownship is initialized and traffic headings are marked as valid.
-	const double lat_uav = 46.52342;
-	const double lon_uav = 6.524234;
-	const float alt_uav = 400.f;
-	const matrix::Vector3f uav_vel{5.f, 10.f, 2.f};
-
-	set_default_uav_state(lat_uav, lon_uav, alt_uav, uav_vel);
-	sync_navigator_topics();
-
-	const uint16_t flags = transponder_report_s::PX4_ADSB_FLAGS_VALID_COORDS |
-			       transponder_report_s::PX4_ADSB_FLAGS_VALID_ALTITUDE |
-			       transponder_report_s::PX4_ADSB_FLAGS_VALID_HEADING |
-			       transponder_report_s::PX4_ADSB_FLAGS_VALID_VELOCITY;
-
-	// WHEN: Traffic is published with non-finite heading values.
-	for (const float heading : {std::numeric_limits<float>::quiet_NaN(), std::numeric_limits<float>::infinity()}) {
-		drain_detect_and_avoid_topic();
-
-		transponder_report_s tr = create_transponder_report(14545057, "DDF0A1", lat_uav, lon_uav,
-					  alt_uav, 30.f, 0.f, flags);
-		tr.heading = heading;
-		publish_transponder_report_and_check(tr);
-
-		// THEN: No detect_and_avoid output is published and no conflict is buffered.
-		EXPECT_FALSE(_detect_and_avoid_sub.update());
-
-		conflict_info_s conflict;
-		navigator->get_detect_and_avoid()->get_most_urgent_conflict(conflict);
-		EXPECT_EQ(conflict.conflict_level, kDaaConflictLvlNone);
-	}
-}
-
-// WHY: A hovering ownship has no meaningful course-over-ground, so crosstrack mode must use the local-position yaw instead of atan2f(0, 0).
-// WHAT: Publish a hovering ownship with unavailable local yaw and verify crosstrack processing fails instead of inventing a north heading.
-TEST_F(DetectAndAvoidTest, HoveringCrosstrackRejectsNonFiniteOwnshipHeading)
-{
-	// GIVEN: DetectAndAvoid is recreated in crosstrack mode.
-	const int crosstrack_standard = detect_and_avoid_s::DAA_STANDARD_CROSSTRACK;
-	param_set(param_handle(px4::params::DAA_STANDARD), &crosstrack_standard);
-	recreate_navigator();
-
-	// GIVEN: Ownship is hovering and local-position yaw is unavailable.
-	const double lat_uav = 46.52342;
-	const double lon_uav = 6.524234;
-	const float alt_uav = 400.f;
-	const matrix::Vector3f uav_vel{0.f, 0.f, 0.f};
-
-	publish_global_pos(lat_uav, lon_uav, alt_uav);
-	publish_land_status(false);
-	publish_vehicle_status(vehicle_status_s::NAVIGATION_STATE_AUTO_MISSION, vehicle_status_s::ARMING_STATE_ARMED);
-	publish_local_pos_vel(uav_vel, hrt_absolute_time(), std::numeric_limits<float>::quiet_NaN());
-	sync_navigator_topics();
-	drain_detect_and_avoid_topic();
-
-	const uint16_t flags = transponder_report_s::PX4_ADSB_FLAGS_VALID_COORDS |
-			       transponder_report_s::PX4_ADSB_FLAGS_VALID_ALTITUDE |
-			       transponder_report_s::PX4_ADSB_FLAGS_VALID_HEADING |
-			       transponder_report_s::PX4_ADSB_FLAGS_VALID_VELOCITY;
-	transponder_report_s tr = create_transponder_report(14545057, "DDF0A1", lat_uav, lon_uav,
-				  alt_uav, 30.f, 0.f, flags);
-	tr.heading = 3.f * M_PI_2_F;
-
-	// WHEN: A traffic report that would otherwise be processed by crosstrack mode is received.
-	publish_transponder_report_and_check(tr);
-
-	// THEN: No output is published and no conflict is buffered because ownship yaw is not finite.
-	EXPECT_FALSE(_detect_and_avoid_sub.update());
-
-	conflict_info_s conflict;
-	navigator->get_detect_and_avoid()->get_most_urgent_conflict(conflict);
-	EXPECT_EQ(conflict.conflict_level, kDaaConflictLvlNone);
 }
 
 // WHY: DAA must stop exposing buffered conflicts once the uav's position or velocity timestamps go stale.
@@ -2381,317 +1584,3 @@ TEST_F(DetectAndAvoidTest, BufferFull)
 	check_highest_conflict(conflict.conflict_level,
 			       navigator->get_detect_and_avoid()->conflict_lvl_requires_action(conflict.conflict_level));
 }
-
-// WHY: Ownship reports must never enter the avoidance buffer regardless of whether identity comes from ICAO, callsign, or UAS ID.
-// WHAT: Configure each self identifier format and verify DetectAndAvoid recognizes and suppresses matching traffic.
-TEST_F(DetectAndAvoidTest, SelfDetection)
-{
-	/*
-		Method used to set call sign parameters:
-			1. Start with callsign
-			2. Convert to uint64:
-				callsign_to_uint64(const char callsign[kCallsignLength]);
-			3. Convert to two uint32
-				const uint32_t lower = static_cast<uint32_t>(unique_id.id & 0xFFFFFFFF);
-				const uint32_t upper = static_cast<uint32_t>((unique_id.id >> 32) & 0xFFFFFFFF);
-
-		Method used to set ICAO parameters:
-			uint32_t icao_uint = static_cast<uint32_t>(std::stoul(icao_str, nullptr, 16));
-
-	*/
-
-	// GIVEN: DetectAndAvoid is active with default parameters.
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->is_activated());
-
-	// GIVEN: Random ADS-B callsigns will be configured as ownship identifiers.
-	char callsign[9];
-	transponder_report_s tr{};
-	unique_id_s unique_id;
-
-	// WHEN: Callsign identifiers are written into the self-identification parameters.
-	for (int i = 0; i < 30; ++i) {
-		generate_random_callsign(callsign, 8);
-		const uint64_t callsign_64 = navigator->get_detect_and_avoid()->callsign_to_uint64(callsign);
-		const uint32_t own_adsb_callsign1 = static_cast<uint32_t>(callsign_64 & 0xFFFFFFFF);
-		const uint32_t own_adsb_callsign2 = static_cast<uint32_t>((callsign_64 >> 32) & 0xFFFFFFFF);
-		param_set(param_handle(px4::params::ADSB_CALLSIGN_1), &own_adsb_callsign1);
-		param_set(param_handle(px4::params::ADSB_CALLSIGN_2), &own_adsb_callsign2);
-
-		navigator->get_detect_and_avoid()->updateParams();
-
-		tr.icao_address = 0; // Invalid ICAO
-		strncpy(&tr.callsign[0], callsign, sizeof(tr.callsign) - 1);
-		tr.callsign[sizeof(tr.callsign) - 1] = 0;
-		tr.flags = transponder_report_s::PX4_ADSB_FLAGS_VALID_CALLSIGN;
-
-		navigator->get_detect_and_avoid()->get_unique_id(tr, unique_id);
-
-		// THEN: Matching callsigns are recognized as self-detections.
-		EXPECT_TRUE(unique_id.encoding == detect_and_avoid_s::UNIQUE_ID_ENCODING_ADSB_CALLSIGN);
-		EXPECT_TRUE(unique_id.id == callsign_64);
-		EXPECT_TRUE(navigator->get_detect_and_avoid()->is_self_detection(unique_id));
-	}
-
-	// WHEN: The primary ICAO address is configured as ownship.
-	const uint32_t own_icao = 6593425;
-	tr.icao_address = own_icao;
-	navigator->get_detect_and_avoid()->get_unique_id(tr, unique_id);
-
-	// THEN: With the default disabled primary ICAO parameter, real traffic is not treated as ownship.
-	EXPECT_FALSE(navigator->get_detect_and_avoid()->is_self_detection(unique_id));
-
-	param_set(param_handle(px4::params::ADSB_ICAO_ID), &own_icao);
-	navigator->get_detect_and_avoid()->updateParams();
-
-	tr.icao_address = own_icao;
-	strncpy(&tr.callsign[0], callsign, sizeof(tr.callsign) - 1);
-	tr.callsign[sizeof(tr.callsign) - 1] = 0;
-	tr.flags = transponder_report_s::PX4_ADSB_FLAGS_VALID_CALLSIGN;
-
-	navigator->get_detect_and_avoid()->get_unique_id(tr, unique_id);
-
-	// THEN: Matching primary ICAO traffic is treated as self-detection.
-	EXPECT_TRUE(unique_id.encoding == detect_and_avoid_s::UNIQUE_ID_ENCODING_ICAO);
-	EXPECT_TRUE(unique_id.id == own_icao);
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->is_self_detection(unique_id));
-
-	// WHEN: The secondary ICAO address is configured as ownship.
-	const uint32_t own_icao_2 = 3318901;
-	param_set(param_handle(px4::params::ADSB_ICAO_ID_2), &own_icao_2);
-	navigator->get_detect_and_avoid()->updateParams();
-
-	tr.icao_address = own_icao_2;
-	strncpy(&tr.callsign[0], callsign, sizeof(tr.callsign) - 1);
-	tr.callsign[sizeof(tr.callsign) - 1] = 0;
-	tr.flags = transponder_report_s::PX4_ADSB_FLAGS_VALID_CALLSIGN;
-
-	navigator->get_detect_and_avoid()->get_unique_id(tr, unique_id);
-
-	// THEN: Matching secondary ICAO traffic is also treated as self-detection.
-	EXPECT_TRUE(unique_id.encoding == detect_and_avoid_s::UNIQUE_ID_ENCODING_ICAO);
-	EXPECT_TRUE(unique_id.id == own_icao_2);
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->is_self_detection(unique_id));
-
-	// WHEN: The traffic report carries only a UAS ID.
-	transponder_report_s tr_uas_id{};
-	tr_uas_id.icao_address = 0;
-	memset(tr_uas_id.callsign, 0, sizeof(tr_uas_id.callsign));
-
-#ifndef BOARD_HAS_NO_UUID
-	px4_guid_t px4_guid {};
-	board_get_px4_guid(px4_guid);
-	memcpy(tr_uas_id.uas_id, px4_guid, sizeof(px4_guid));
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->get_unique_id(tr_uas_id, unique_id));
-
-	// THEN: On boards with a UUID, the local UAS ID is recognized as self.
-	EXPECT_TRUE(unique_id.encoding == detect_and_avoid_s::UNIQUE_ID_ENCODING_UAS_ID);
-	EXPECT_TRUE(unique_id.id == navigator->get_detect_and_avoid()->last_uas_id_bytes_to_uint64(tr_uas_id.uas_id));
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->is_self_detection(unique_id));
-#else
-
-	for (int i = 0; i < PX4_GUID_BYTE_LENGTH; ++i) {
-		tr_uas_id.uas_id[i] = 0xe0 + i;
-	}
-
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->get_unique_id(tr_uas_id, unique_id));
-
-	// THEN: Without a board UUID, an arbitrary UAS ID is not considered self.
-	EXPECT_TRUE(unique_id.encoding == detect_and_avoid_s::UNIQUE_ID_ENCODING_UAS_ID);
-	EXPECT_FALSE(navigator->get_detect_and_avoid()->is_self_detection(unique_id));
-#endif
-
-	// WHEN: Traffic provides no usable identity fields.
-	transponder_report_s tr_no_id{};
-	tr_no_id.icao_address = 0;
-	strncpy(&tr_no_id.callsign[0], callsign, sizeof(tr_no_id.callsign) - 1);
-	tr_no_id.callsign[sizeof(tr.callsign) - 1] = 0;
-	tr_no_id.flags = 0;
-
-	// THEN: No unique identifier can be extracted.
-	EXPECT_FALSE(navigator->get_detect_and_avoid()->get_unique_id(tr_no_id, unique_id));
-}
-
-// WHY: Callsign packing and unpacking must round-trip cleanly so buffer keys and operator-facing messages stay consistent.
-// WHAT: Convert random callsigns to uint64 and back, then verify the recovered string produces the same encoded value and reject unterminated input.
-TEST_F(DetectAndAvoidTest, ReversibleCallsign)
-{
-	/*
-		Ensure conversion is reversible
-			Callsign:
-				- uint = callsign_to_uint64
-				- recovered = convert_uint64_callsign_to_str(uint)
-				- back_to_uint = callsign_to_uint64(recovered)
-
-				--> uint == back_to_uint
-	*/
-
-	// GIVEN: DetectAndAvoid is active and the callsign helpers are available.
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->is_activated());
-
-	char callsign[9];
-
-	// WHEN: Random null-terminated callsigns are converted to uint64 and back.
-	for (int i = 0; i < 30; ++i) {
-		generate_random_callsign(callsign, 8);
-		const uint64_t callsign_uint = navigator->get_detect_and_avoid()->callsign_to_uint64(callsign);
-		char recovered[9];
-		navigator->get_detect_and_avoid()->convert_uint64_callsign_to_str(callsign_uint, recovered);
-		const uint64_t final_uint = navigator->get_detect_and_avoid()->callsign_to_uint64(recovered);
-
-		// THEN: The encoding round-trips without changing the packed value.
-		EXPECT_TRUE(callsign_uint == final_uint);
-
-		// Only debug print once
-		if (i == 0) {
-
-			PX4_DEBUG("TEST_F: UniqueID, init: %s (int %lu), recovered: %s (int %lu)", callsign, callsign_uint, recovered,
-				  final_uint);
-
-			char part1_recovered[9];
-			char part2_recovered[9];
-			uint32_t lower = static_cast<uint32_t>(callsign_uint & 0xFFFFFFFF);
-			uint32_t upper = static_cast<uint32_t>((callsign_uint >> 32) & 0xFFFFFFFF);
-			navigator->get_detect_and_avoid()->convert_uint64_callsign_to_str(uint64_t(lower), part1_recovered);
-			navigator->get_detect_and_avoid()->convert_uint64_callsign_to_str(uint64_t(upper), part2_recovered);
-			PX4_DEBUG("TEST_F: full: %s =? %s---%s == PART1---PART2", recovered, part1_recovered, part2_recovered);
-		}
-	}
-
-	// WHEN: A callsign is not null terminated.
-	char callsign_non_null[9];
-	generate_random_callsign(callsign_non_null, 8);
-	callsign_non_null[8] = 1;
-
-	// THEN: The helper rejects the malformed input.
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->callsign_to_uint64(callsign_non_null) == 0);
-}
-
-// WHY: Callsigns are stored as bytes, so high-bit characters must only affect their own byte slot and never sign-extend across the uint64_t key.
-// WHAT: Encode a callsign containing bytes with the MSB set and verify the packed value preserves only those byte values.
-TEST_F(DetectAndAvoidTest, CallsignPackingDoesNotSignExtend)
-{
-	const char callsign[kCallsignLength] {static_cast<char>(0x80), static_cast<char>(0xFF), '\0'};
-	const uint64_t callsign_uint = navigator->get_detect_and_avoid()->callsign_to_uint64(callsign);
-
-	EXPECT_EQ(callsign_uint, static_cast<uint64_t>(0xFF80u));
-}
-
-// WHY: MAVLink fixed-length string fields copy every byte, so unused callsign bytes must be deterministic after decoding.
-// WHAT: Decode a short callsign into a pre-filled buffer and verify all bytes after the terminator are cleared.
-TEST_F(DetectAndAvoidTest, CallsignDecodeClearsUnusedBytes)
-{
-	const char callsign[kCallsignLength] {'A', 'B', '\0'};
-	const uint64_t callsign_uint = navigator->get_detect_and_avoid()->callsign_to_uint64(callsign);
-	char recovered[kCallsignLength];
-	memset(recovered, 0x7F, sizeof(recovered));
-
-	navigator->get_detect_and_avoid()->convert_uint64_callsign_to_str(callsign_uint, recovered);
-
-	EXPECT_STREQ(recovered, "AB");
-
-	for (int i = 2; i < kCallsignLength; ++i) {
-		EXPECT_EQ(recovered[i], '\0');
-	}
-}
-
-// WHY: ICAO identifiers are 24-bit values, so operator-facing formatting must stay fixed-width and avoid platform-specific truncation.
-// WHAT: Format representative ICAO values and verify the output is uppercase, zero-padded, and derived from the low 24 bits.
-TEST_F(DetectAndAvoidTest, FormatsIcaoAsFixedWidthHex)
-{
-	// GIVEN: A buffer sized for a 6-character ICAO address plus the null terminator.
-	char icao_buffer[kIcaoLength] {};
-
-	// WHEN: A short ICAO value is formatted.
-	DetectAndAvoid::convert_icao_uint32_to_hex_str(0xABCu, icao_buffer, sizeof(icao_buffer));
-
-	// THEN: The helper zero-pads and uppercases the address.
-	EXPECT_STREQ(icao_buffer, "000ABC");
-
-	// WHEN: High bits are present in the input value.
-	DetectAndAvoid::convert_icao_uint32_to_hex_str(0x12ABCDEFu, icao_buffer, sizeof(icao_buffer));
-
-	// THEN: Only the ICAO address bits are rendered.
-	EXPECT_STREQ(icao_buffer, "ABCDEF");
-}
-
-// WHY: Reduced UAS-ID packing must preserve the exact GUID tail bytes used as the unique identifier.
-// WHAT: Round-trip random and zero-valued GUID tails through the uint64 conversion helpers and compare every byte.
-TEST_F(DetectAndAvoidTest, ReversibleUasId)
-{
-	/*
-		Ensure conversion is reversible
-			UAS_ID:
-				- uint = last_uas_id_bytes_to_uint64(uas_id);
-				- recovered = uint64_to_last_uas_id_bytes(uint);
-
-				--> last_uas_id_bytes == recovered_bytes
-	*/
-
-	// GIVEN: DetectAndAvoid is active and reduced UAS-ID packing is available.
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->is_activated());
-
-	uint8_t uas_id[PX4_GUID_BYTE_LENGTH];
-
-	// WHEN: Random UAS IDs are packed into uint64 and unpacked again.
-	for (int i = 0; i < 30; ++i) {
-		generate_random_UAS_ID(uas_id);
-		const uint64_t uas_id_uint = navigator->get_detect_and_avoid()->last_uas_id_bytes_to_uint64(
-						     uas_id); // Converts last kIdEncodingNbBytes bytes
-		uint8_t recovered[kIdEncodingNbBytes]; // Gets the last kIdEncodingNbBytes back
-		navigator->get_detect_and_avoid()->uint64_to_last_uas_id_bytes(uas_id_uint, recovered);
-
-		// THEN: The packed representation preserves the trailing GUID bytes exactly.
-		for (int k = 0; k < kIdEncodingNbBytes; ++k) {
-			EXPECT_TRUE(uas_id[PX4_GUID_BYTE_LENGTH - kIdEncodingNbBytes + k] == recovered[k]);
-
-			if (i == 0) {
-				PX4_DEBUG(" Byte %d/%d: uas == %d == %d == recovered", PX4_GUID_BYTE_LENGTH - kIdEncodingNbBytes + k + 1,
-					  PX4_GUID_BYTE_LENGTH, uas_id[PX4_GUID_BYTE_LENGTH - kIdEncodingNbBytes + k], recovered[k]);
-			}
-
-		}
-	}
-
-	// WHEN: The UAS ID is all zeros.
-	uint8_t zero_uas_id[PX4_GUID_BYTE_LENGTH];
-
-	for (int i = 0; i < PX4_GUID_BYTE_LENGTH; i++) {
-		zero_uas_id[i] = 0;
-	}
-
-	const uint64_t zero_uas_id_uint = navigator->get_detect_and_avoid()->last_uas_id_bytes_to_uint64(
-			zero_uas_id); // Converts last kIdEncodingNbBytes bytes
-
-	ASSERT_TRUE(zero_uas_id_uint == 0);
-
-	// WHEN: The zero value is unpacked again.
-	uint8_t zero_recovered[kIdEncodingNbBytes];
-	navigator->get_detect_and_avoid()->uint64_to_last_uas_id_bytes(zero_uas_id_uint, zero_recovered);
-
-	// THEN: The unpacked trailing bytes remain zeroed.
-	for (int k = 0; k < kIdEncodingNbBytes; ++k) {
-		EXPECT_TRUE(0 == zero_recovered[k]);
-	}
-}
-
-// WHY: Automatic DAA actions must only escalate when they are stronger than the current navigator state permits.
-// WHAT: Evaluate each requested action across representative navigation states and verify the expected escalation decisions.
-TEST_P(DetectAndAvoidActionPriorityTest, ActionPriorities)
-{
-	EXPECT_TRUE(navigator->get_detect_and_avoid()->is_activated());
-
-	const action_priority_test_case_s &test_case = GetParam();
-	check_eval_conflict_escalation_action(navigator.get(), test_case.expected_outcome, test_case.action);
-}
-
-INSTANTIATE_TEST_SUITE_P(
-	AllActions,
-	DetectAndAvoidActionPriorityTest,
-	::testing::Values(
-		action_priority_test_case_s{"Disabled", DaaAction::kDisabled, {false, false, false, false}},
-		action_priority_test_case_s{"WarnOnly", DaaAction::kWarnOnly, {false, false, false, false}},
-		action_priority_test_case_s{"Hold", DaaAction::kPositionHoldMode, {true, false, false, false}},
-		action_priority_test_case_s{"ReturnMode", DaaAction::kReturnMode, {true, true, false, false}},
-		action_priority_test_case_s{"LandMode", DaaAction::kLandMode, {true, true, true, false}},
-		action_priority_test_case_s{"Terminate", DaaAction::kTerminate, {true, true, true, true}}),
-	action_priority_test_case_name);
