+++ test/CMakeLists.txt
@@ -15,9 +15,13 @@ else()
     set(TESTFILTERARG "")
 endif()
 
+# Protect regex alternations (e.g., "a|b") from being split by the shell.
+# New variable so px4_add_gtest.cmake can still use the raw string.
+string(REPLACE "|" "\\|" TESTFILTER_SHELL "${TESTFILTER}")
+
 add_custom_target(test_results
         # antlr4_tests_NOT_BUILT gets added by fuzztest
-        COMMAND GTEST_COLOR=1 ${CMAKE_CTEST_COMMAND} --output-on-failure -T Test ${TESTFILTERARG} ${TESTFILTER} --exclude-regex "antlr4_tests_NOT_BUILT"
+        COMMAND GTEST_COLOR=1 ${CMAKE_CTEST_COMMAND} --output-on-failure -T Test ${TESTFILTERARG} ${TESTFILTER_SHELL} --exclude-regex "antlr4_tests_NOT_BUILT"
         DEPENDS
         px4
         examples__dyn_hello
